Add S3 Connection Attributes to DMS Endpoint
Adds a few commonly used parameters to the explicit S3 connection block.

Please note that IMO, we should not default those settings because they
are defaulted on AWS side. However, the API is a bit inconsistent in
which properties the API returns on GET if the resource was not
_created_ with particular resources (i.e.
`parquet_timestamp_in_millisecond` is always returned, `data_format` is
not if it was not explicitly used for creation, yet it defaults to
`csv`.

In order to make tests and usage of the provider a bit more sane, I
defaulted all properties.